chore: misc. clean up
