+++ internal/handler/authentication.go
@@ -14,10 +14,41 @@ import (
 	"gorm.io/gorm"
 )
 
-type AuthenticationMiddleware struct{ c config.Config }
+func NewAuthentication(c config.Config, signInService signInService) AuthenticationMiddleware {
+	return AuthenticationMiddleware{c, signInService}
+}
 
-func NewAuthentication(c config.Config) *AuthenticationMiddleware {
-	return &AuthenticationMiddleware{c}
+type signInService interface {
+	SignIn(email string, password string) (*model.User, error)
+}
+
+type AuthenticationMiddleware struct {
+	c             config.Config
+	signInService signInService
+}
+
+// BasicAuthentication Inspiration: https://www.pandurang-waghulde.com/custom-http-basic-authentication-using-gin/
+func (m AuthenticationMiddleware) BasicAuthentication(c *gin.Context) {
+	username, password, ok := c.Request.BasicAuth()
+	if !ok {
+		m.handleError(c, errors.New("invalid Authorization header format"))
+		return
+	}
+
+	u, err := m.signInService.SignIn(username, password)
+	if err != nil {
+		m.handleError(c, err)
+		return
+	}
+
+	c.Set("user", u)
+	c.Next()
+}
+
+func (m AuthenticationMiddleware) handleError(c *gin.Context, e error) {
+	// Trigger username/password prompt
+	c.Header("WWW-Authenticate", "Basic realm=\"DHIS2\"")
+	_ = c.AbortWithError(http.StatusUnauthorized, e)
 }
 
 func (m AuthenticationMiddleware) TokenAuthentication(c *gin.Context) {
@@ -66,8 +97,8 @@ func parseRequest(request *http.Request, key *rsa.PublicKey) (*model.User, error
 	return extractUser(userData)
 }
 
-func extractUser(userData interface{}) (*model.User, error) {
-	userMap, ok := userData.(map[string]interface{})
+func extractUser(userData any) (*model.User, error) {
+	userMap, ok := userData.(map[string]any)
 	if !ok {
 		return nil, errors.New("failed to parse user data")
 	}
@@ -84,12 +115,12 @@ func extractUser(userData interface{}) (*model.User, error) {
 	return user, nil
 }
 
-func extractGroups(key string, userMap map[string]interface{}) []model.Group {
-	groupsData, ok := userMap[key].([]interface{})
+func extractGroups(key string, userMap map[string]any) []model.Group {
+	groupsData, ok := userMap[key].([]any)
 	if ok {
 		groups := make([]model.Group, len(groupsData))
 		for i := 0; i < len(groupsData); i++ {
-			group := groupsData[i].(map[string]interface{})
+			group := groupsData[i].(map[string]any)
 			groups[i] = model.Group{
 				Name:     group["Name"].(string),
 				Hostname: group["Hostname"].(string),
