+++ pkg/database/routes.go
@@ -1,11 +1,11 @@
 package database
 
 import (
-	"github.com/dhis2-sre/im-manager/internal/middleware"
+	"github.com/dhis2-sre/im-manager/internal/handler"
 	"github.com/gin-gonic/gin"
 )
 
-func Routes(router *gin.Engine, authenticationMiddleware middleware.AuthenticationMiddleware, handler Handler) {
+func Routes(router *gin.Engine, authenticationMiddleware handler.AuthenticationMiddleware, handler Handler) {
 	router.GET("/databases/external/:uuid", handler.ExternalDownload)
 
 	tokenAuthenticationRouter := router.Group("/databases")
