+++ pkg/group/router.go
@@ -1,11 +1,12 @@
 package group
 
 import (
+	"github.com/dhis2-sre/im-manager/internal/handler"
 	"github.com/dhis2-sre/im-manager/internal/middleware"
 	"github.com/gin-gonic/gin"
 )
 
-func Routes(r *gin.Engine, authenticationMiddleware middleware.AuthenticationMiddleware, authorizationMiddleware middleware.AuthorizationMiddleware, handler Handler) {
+func Routes(r *gin.Engine, authenticationMiddleware handler.AuthenticationMiddleware, authorizationMiddleware middleware.AuthorizationMiddleware, handler Handler) {
 	tokenAuthenticationRouter := r.Group("")
 	tokenAuthenticationRouter.Use(authenticationMiddleware.TokenAuthentication)
