+++ pkg/instance/router.go
@@ -1,11 +1,11 @@
 package instance
 
 import (
-	"github.com/dhis2-sre/im-manager/internal/middleware"
+	"github.com/dhis2-sre/im-manager/internal/handler"
 	"github.com/gin-gonic/gin"
 )
 
-func Routes(r *gin.Engine, authenticationMiddleware middleware.AuthenticationMiddleware, handler Handler) {
+func Routes(r *gin.Engine, authenticationMiddleware handler.AuthenticationMiddleware, handler Handler) {
 	tokenAuthenticationRouter := r.Group("")
 	tokenAuthenticationRouter.Use(authenticationMiddleware.TokenAuthentication)
